+++ api/utils.py
@@ -12,64 +12,68 @@ logger = logging.getLogger(__name__)
 def create_gmw(
     bro_id: str, metadata: dict, sourcedocument_data: dict, data_owner: str
 ) -> None:
-    gmw = GMW.objects.create(
+    gmw = GMW.objects.update_or_create(
         bro_id=bro_id,
         data_owner=data_owner,
-        internal_id=sourcedocument_data.get("objectIdAccountableParty"),
-        delivery_accountable_party=metadata.get("deliveryAccountableParty"),
-        quality_regime=metadata.get("qualityRegime"),
-        well_construction_date=sourcedocument_data.get("wellConstructionDate"),
-        delivery_context=sourcedocument_data.get("deliveryContext"),
-        construction_standard=sourcedocument_data.get("constructionStandard"),
-        initial_function=sourcedocument_data.get("initialFunction"),
-        ground_level_stable=sourcedocument_data.get("groundLevelStable"),
-        well_stability=sourcedocument_data.get("wellStability"),
-        nitg_code=sourcedocument_data.get("nitgCode"),
-        well_code=sourcedocument_data.get("wellCode"),
-        owner=sourcedocument_data.get("owner"),
-        well_head_protector=sourcedocument_data.get("wellHeadProtector"),
-        delivered_location=sourcedocument_data.get("deliveredLocation"),
-        horizontal_positioning_method=sourcedocument_data.get(
-            "horizontalPositioningMethod"
-        ),
-        local_vertical_reference_point=sourcedocument_data.get(
-            "localVerticalReferencePoint"
-        ),
-        offset=sourcedocument_data.get("offset"),
-        vertical_datum=sourcedocument_data.get("verticalDatum"),
-        ground_level_position=sourcedocument_data.get("groundLevelPosition"),
-        ground_level_positioning_method=sourcedocument_data.get(
-            "groundLevelPositioningMethod"
-        ),
-        standardized_location=sourcedocument_data.get("standardizedLocation"),
-    )
+        defaults={
+            "internal_id": sourcedocument_data.get("objectIdAccountableParty"),
+            "delivery_accountable_party": metadata.get("deliveryAccountableParty"),
+            "quality_regime": metadata.get("qualityRegime"),
+            "well_construction_date": sourcedocument_data.get("wellConstructionDate"),
+            "delivery_context": sourcedocument_data.get("deliveryContext"),
+            "construction_standard": sourcedocument_data.get("constructionStandard"),
+            "initial_function": sourcedocument_data.get("initialFunction"),
+            "ground_level_stable": sourcedocument_data.get("groundLevelStable"),
+            "well_stability": sourcedocument_data.get("wellStability"),
+            "nitg_code": sourcedocument_data.get("nitgCode"),
+            "well_code": sourcedocument_data.get("wellCode"),
+            "owner": sourcedocument_data.get("owner"),
+            "well_head_protector": sourcedocument_data.get("wellHeadProtector"),
+            "delivered_location": sourcedocument_data.get("deliveredLocation"),
+            "horizontal_positioning_method": sourcedocument_data.get(
+                "horizontalPositioningMethod"
+            ),
+            "local_vertical_reference_point": sourcedocument_data.get(
+                "localVerticalReferencePoint"
+            ),
+            "offset": sourcedocument_data.get("offset"),
+            "vertical_datum": sourcedocument_data.get("verticalDatum"),
+            "ground_level_position": sourcedocument_data.get("groundLevelPosition"),
+            "ground_level_positioning_method": sourcedocument_data.get(
+                "groundLevelPositioningMethod"
+            ),
+            "standardized_location": sourcedocument_data.get("standardizedLocation"),
+        },
+    )[0]
     for tube in sourcedocument_data.get("monitoringTubes", []):
-        MonitoringTube.objects.create(
+        MonitoringTube.objects.update_or_create(
             gmw=gmw,
             data_owner=gmw.data_owner,
-            tube_number=tube.get("tubeNumber"),
-            tube_type=tube.get("tubeType"),
-            artesian_well_cap_present=tube.get("artesianWellCapPresent"),
-            sediment_sump_present=tube.get("sedimentSumpPresent"),
-            sediment_sump_length=tube.get("sedimentSumpLength"),
-            number_of_geo_ohm_cables=tube.get("numberOfGeoOhmCables", 0),
-            geo_ohm_cables=tube.get("geoOhmCables", []),
-            tube_top_diameter=tube.get("tubeTopDiameter"),
-            variable_diameter=tube.get("variableDiameter"),
-            tube_status=tube.get("tubeStatus"),
-            tube_top_position=tube.get("tubeTopPosition"),
-            tube_top_positioning_method=tube.get("tubeTopPositioningMethod"),
-            tube_part_inserted=tube.get("tubePartInserted"),
-            tube_in_use=tube.get("tubeInUse"),
-            tube_packing_material=tube.get("tubePackingMaterial"),
-            tube_material=tube.get("tubeMaterial"),
-            glue=tube.get("glue"),
-            screen_length=tube.get("screenLength"),
-            screen_protection=tube.get("screenProtection"),
-            sock_material=tube.get("sockMaterial"),
-            screen_top_position=tube.get("screenTopPosition"),
-            screen_bottom_position=tube.get("screenBottomPosition"),
-            plain_tube_part_length=tube.get("plainTubePartLength"),
+            defaults={
+                "tube_number": tube.get("tubeNumber"),
+                "tube_type": tube.get("tubeType"),
+                "artesian_well_cap_present": tube.get("artesianWellCapPresent"),
+                "sediment_sump_present": tube.get("sedimentSumpPresent"),
+                "sediment_sump_length": tube.get("sedimentSumpLength"),
+                "number_of_geo_ohm_cables": tube.get("numberOfGeoOhmCables", 0),
+                "geo_ohm_cables": tube.get("geoOhmCables", []),
+                "tube_top_diameter": tube.get("tubeTopDiameter"),
+                "variable_diameter": tube.get("variableDiameter"),
+                "tube_status": tube.get("tubeStatus"),
+                "tube_top_position": tube.get("tubeTopPosition"),
+                "tube_top_positioning_method": tube.get("tubeTopPositioningMethod"),
+                "tube_part_inserted": tube.get("tubePartInserted"),
+                "tube_in_use": tube.get("tubeInUse"),
+                "tube_packing_material": tube.get("tubePackingMaterial"),
+                "tube_material": tube.get("tubeMaterial"),
+                "glue": tube.get("glue"),
+                "screen_length": tube.get("screenLength"),
+                "screen_protection": tube.get("screenProtection"),
+                "sock_material": tube.get("sockMaterial"),
+                "screen_top_position": tube.get("screenTopPosition"),
+                "screen_bottom_position": tube.get("screenBottomPosition"),
+                "plain_tube_part_length": tube.get("plainTubePartLength"),
+            },
         )
 
     logger.info(f"Sucessfully created {gmw} with {gmw.tubes.count()} monitoring tubes.")
@@ -91,13 +95,15 @@ def create_gmw_event(
         logger.error(f"GMW not found for bro_id={bro_id}, owner={data_owner}")
         return
 
-    Event.objects.create(
+    Event.objects.update_or_create(
         gmw=gmw,
         event_name=event_type,
         event_date=sourcedocument_data.get("eventDate"),
-        metadata=metadata,
-        sourcedocument_data=sourcedocument_data,
         data_owner=data_owner,
+        defaults={
+            "metadata": metadata,
+            "sourcedocument_data": sourcedocument_data,
+        },
     )
 
 
@@ -115,15 +121,17 @@ def find_linked_gmns(gmn_bro_ids: list[str] | str) -> list[GMN]:
 def create_gld(
     bro_id: str, metadata: dict, sourcedocument_data: dict, data_owner: str
 ) -> None:
-    GLD.objects.create(
-        internal_id=sourcedocument_data.get("objectIdAccountableParty"),
+    GLD.objects.update_or_create(
         bro_id=bro_id,
-        delivery_accountable_party=metadata.get("deliveryAccountableParty"),
-        linked_gmns=sourcedocument_data.get("linkedGmns", []),
-        quality_regime=metadata.get("qualityRegime"),
-        gmw_bro_id=sourcedocument_data.get("gmwBroId"),
-        tube_number=sourcedocument_data.get("tubeNumber"),
         data_owner=data_owner,
+        defaults={
+            "internal_id": sourcedocument_data.get("objectIdAccountableParty"),
+            "delivery_accountable_party": metadata.get("deliveryAccountableParty"),
+            "linked_gmns": sourcedocument_data.get("linkedGmns", []),
+            "quality_regime": metadata.get("qualityRegime"),
+            "gmw_bro_id": sourcedocument_data.get("gmwBroId"),
+            "tube_number": sourcedocument_data.get("tubeNumber"),
+        },
     )
     return
 
@@ -131,16 +139,18 @@ def create_gld(
 def create_gmn(
     bro_id: str, metadata: dict, sourcedocument_data: dict, data_owner: str
 ) -> None:
-    GMN.objects.create(
+    GMN.objects.update_or_create(
         bro_id=bro_id,
         data_owner=data_owner,
-        internal_id=sourcedocument_data.get("objectIdAccountableParty"),
-        quality_regime=metadata.get("qualityRegime"),
-        name=sourcedocument_data.get("name"),
-        delivery_context=sourcedocument_data.get("deliveryContext"),
-        monitoring_purpose=sourcedocument_data.get("monitoringPurpose"),
-        groundwater_aspect=sourcedocument_data.get("groundwaterAspect"),
-        start_date_monitoring=sourcedocument_data.get("startDateMonitoring"),
+        defaults={
+            "internal_id": sourcedocument_data.get("objectIdAccountableParty"),
+            "quality_regime": metadata.get("qualityRegime"),
+            "name": sourcedocument_data.get("name"),
+            "delivery_context": sourcedocument_data.get("deliveryContext"),
+            "monitoring_purpose": sourcedocument_data.get("monitoringPurpose"),
+            "groundwater_aspect": sourcedocument_data.get("groundwaterAspect"),
+            "start_date_monitoring": sourcedocument_data.get("startDateMonitoring"),
+        },
     )
     return
 
@@ -148,14 +158,18 @@ def create_gmn(
 def create_frd(
     bro_id: str, metadata: dict, sourcedocument_data: dict, data_owner: str
 ) -> None:
-    FRD.objects.create(
+    FRD.objects.update_or_create(
         bro_id=bro_id,
         data_owner=data_owner,
-        internal_id=sourcedocument_data.get("objectIdAccountableParty"),
-        delivery_accountable_party=sourcedocument_data.get("deliveryAccountableParty"),
-        quality_regime=metadata.get("qualityRegime"),
-        gmw_bro_id=sourcedocument_data.get("gmwBroId"),
-        tube_number=sourcedocument_data.get("tubeNumber"),
+        defaults={
+            "internal_id": sourcedocument_data.get("objectIdAccountableParty"),
+            "delivery_accountable_party": sourcedocument_data.get(
+                "deliveryAccountableParty"
+            ),
+            "quality_regime": metadata.get("qualityRegime"),
+            "gmw_bro_id": sourcedocument_data.get("gmwBroId"),
+            "tube_number": sourcedocument_data.get("tubeNumber"),
+        },
     )
     return